+++ cmd/collector/app/server/http.go
@@ -17,7 +17,7 @@ import (
 	"github.com/jaegertracing/jaeger/internal/recoveryhandler"
 	samplinghttp "github.com/jaegertracing/jaeger/internal/sampling/http"
 	"github.com/jaegertracing/jaeger/internal/sampling/samplingstrategy"
-	"github.com/jaegertracing/jaeger/pkg/httpmetrics"
+	"github.com/jaegertracing/jaeger/internal/httpmetrics"
 	"github.com/jaegertracing/jaeger/pkg/metrics"
 	"github.com/jaegertracing/jaeger/pkg/telemetry"
 )
